requirements.txt
discord.py==1.6.0
neispy==3.2.2